requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2024.6.2
charset-normalizer==3.3.2
Django==5.0
django-admin-interface==0.28.1
django-colorfield==0.11.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-restframework==0.0.1
django-tenants==3.6.1
django-wkhtmltopdf==3.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.0
idna==3.7
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
Pillow==10.1.0
psycopg2==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.0
python-slugify==8.0.1
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
requests==2.32.3
rpds-py==0.16.2
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
tzdata==2023.3
Unidecode==1.3.7
uritemplate==4.1.1
urllib3==2.2.2
whitenoise==6.6.0
